Added Content, Image files, for most of the main pages
I have completed the Main homepage, the about me page, the explore page, and the characters page. I have yet to complete the content for the staterra.html page but will continue to do so for personal use in the near future. This will also included detailed bios for character specific pages and region specific pages for worldbuilding.